+++ src/common/utils/deployTips.ts
@@ -117,11 +117,13 @@ You may also have a look to command sfdx hardis:project:clean:references
           name: 'custom-field-not-found',
           label: 'Custom field not found',
           expressionRegex: [/In field: field - no CustomField named (.*) found/gm],
-          tip: `A reference to a custom object is not found:
+          tip: `A reference to a custom field is not found:
 - If you renamed the custom field, do a search/replace in sources with previous field name and new field name
 - If you deleted the custom field, or if you don't want to deploy it, do a search on the custom field name, and remove XML elements referencing it
-- If the object should exist, make sure it is in force-app/main/default/objects/YOUROBJECT/fields and that the field name is in manifest/package.xml in CustomField section
-You may also have a look to command sfdx hardis:project:clean:references
+- If the custom field should exist, make sure it is in force-app/main/default/objects/YOUROBJECT/fields and that the field name is in manifest/package.xml in CustomField section
+- If the field is standard, the error is because the field not available in the org you are trying to deploy to. You can:
+  - Remove the reference to the standard field ( maybe sfdx hardis:project:clean:references can clean auotmatically for you ! )
+  - Activate the required features/license in the target org
 `
         },
         {